feat(kr): decouples add new checkmark button
